Added proper login token support, rearranged the files
