+++ dotnet-music-app/Services/Interfaces/IAlbumService.cs
