+++ dotnet-music-app/Services/Interfaces/IAuthService.cs
