+++ dotnet-music-app/Services/Interfaces/IDbService.cs
