+++ dotnet-music-app/Services/Interfaces/IPlaylistService.cs
