+++ dotnet-music-app/Services/Interfaces/IUserService.cs
