+++ src/features/auth/SignupPage.vue
@@ -1,31 +1,40 @@
 <script setup lang="ts">
 import router from '@/router'
 import { ref } from 'vue'
+import { useAuthStore } from '@/features/auth/authStore'
 
+const authStore = useAuthStore()
+
+const rememberMe = ref(false)
+const display_name = ref('')
 const username = ref('')
 const password = ref('')
 
-const handleLogin = async () => {
+const handleSignup = async () => {
   if (!username.value || !password.value) {
     alert('Please fill in both username and password.')
     return
   }
   try {
-    const response = await fetch('http://localhost:5091/Auth/login', {
+    const response = await fetch('http://localhost:5091/User', {
       method: 'POST',
       headers: {
         'Content-Type': 'application/json'
       },
       body: JSON.stringify({
+        displayName: display_name.value,
         username: username.value,
-        password: password.value
+        password: password.value,
       })
     })
 
-    if (!response.ok) throw new Error('Login failed')
+    if (!response.ok) throw new Error('Sign up failed')
 
     const data = await response.json()
-    console.log('Login success:', data)
+
+    authStore.setToken(data.token, rememberMe.value)
+
+    console.log('Sign up success:', data)
     router.push("/");
   } catch (error) {
     console.error('', error)
@@ -44,6 +53,18 @@ async function toSHA256(message: string): Promise<string> {
 </script>
 
 <template>
+  <div class="flex justify-center">
+    <p class="text-white font-arial mt-5 text-2xl">Display Name:</p>
+  </div>
+  <div class="flex justify-center">
+    <input
+      v-model="display_name"
+      class="form-control me-2 input-rounded text-[#efd0d0] bg-custom max-w-[30vh]"
+      type="text"
+      placeholder="Display Name"
+    />
+  </div>
+
   <div class="flex justify-center">
     <p class="text-white font-arial mt-5 text-2xl">Username:</p>
   </div>
@@ -70,25 +91,25 @@ async function toSHA256(message: string): Promise<string> {
   </div>
 
   <div class="flex justify-center mt-5">
-    <input type="checkbox" id="checkbox" />
+    <input type="checkbox" v-model="rememberMe" />
     <div class="text-white font-arial ml-1">Remember me</div>
   </div>
 
   <div class="flex justify-center mt-5">
     <button
-      @click="handleLogin"
+      @click="handleSignup"
       class="text-white font-arial text-2xl font-semibold bg-[#362323] rounded-xl px-4 py-2"
     >
-      Login
+      Sign up
     </button>
   </div>
   
   <div class="flex justify-center mt-5">
     <RouterLink 
-      to="/signup"
+      to="/login"
       class="text-white font-arial text-xl font-semibold bg-[#362323] rounded-xl px-2 py-1 no-underline"
     >
-      Sign up instead
+      Log in instead
     </RouterLink>
   </div>
 </template>
