+++ src/features/search/SearchPage.vue
@@ -1,5 +1,4 @@
 <script setup lang="ts">
-import MusicPlayer from './MusicPlayer.vue';
 </script>
 
 <template>
