+++ src/types/Album.ts
